Made Program.cs not static anymore and put the console close event handler to a separate class
